update: refactor FastZeroDCE implementation
