+++ docs/models/zero_dce_model.md
@@ -1,1 +1,9 @@
-::: restorers.model.zero_dce.zero_dce_model
+::: restorers.model.zero_dce.zero_dce
+
+::: restorers.model.zero_dce.fast_zero_dce
+
+# Zero-DCE Blocks
+
+::: restorers.model.zero_dce.dw_conv
+
+::: restorers.model.zero_dce.dce_layer
