+++ restorers/model/zero_dce/__init__.py
@@ -1,2 +1,3 @@
 from .dce_layer import DeepCurveEstimationLayer, FastDeepCurveEstimationLayer
-from .zero_dce_model import ZeroDCE, FastZeroDce
+from .zero_dce import ZeroDCE
+from fast_zero_dce import FastZeroDce
